virtcontainers: hypervisor: Move endpoints under hypervisor
There is currently a cyclic logical dependency between network, endpoint
and hypervisor structures, types and APIs. Hypervisor interface
implementations call into the Endpoint interface, which itself calls
into the Network* APIs. Since we want to move all hypervisor implementations
under the hypervisor package, we need to:

- Split the network implementation into 3 pieces: Types definitions,
  core namespace API and Endpoints.
- Move the Endpoint handling under the hypervisor package. This could be
  splitted further if we manage to decouple the Hypervisor interface
  from the Endpoint one, and vice-versa.

Fixes: #1229
