+++ virtcontainers/hack/virtc/main.go
@@ -171,7 +171,7 @@ func buildSandboxConfig(context *cli.Context) (vc.SandboxConfig, error) {
 		return vc.SandboxConfig{}, err
 	}
 
-	netConfig := vc.NetworkConfig{}
+	netConfig := types.NetworkConfig{}
 
 	switch *agentType {
 	case vc.HyperstartAgent:
