+++ virtcontainers/hypervisor/vhostuser_endpoint.go
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-package virtcontainers
+package hypervisor
 
 import (
 	"encoding/hex"
@@ -11,7 +11,7 @@ import (
 	"os"
 
 	"github.com/kata-containers/runtime/virtcontainers/device/config"
-	"github.com/kata-containers/runtime/virtcontainers/hypervisor"
+	"github.com/kata-containers/runtime/virtcontainers/types"
 	"github.com/kata-containers/runtime/virtcontainers/utils"
 )
 
@@ -28,13 +28,13 @@ type VhostUserEndpoint struct {
 	// MAC address of the interface
 	HardAddr           string
 	IfaceName          string
-	EndpointProperties NetworkInfo
+	EndpointProperties types.NetworkInfo
 	EndpointType       EndpointType
 	PCIAddr            string
 }
 
 // Properties returns the properties of the interface.
-func (endpoint *VhostUserEndpoint) Properties() NetworkInfo {
+func (endpoint *VhostUserEndpoint) Properties() types.NetworkInfo {
 	return endpoint.EndpointProperties
 }
 
@@ -54,7 +54,7 @@ func (endpoint *VhostUserEndpoint) Type() EndpointType {
 }
 
 // SetProperties sets the properties of the endpoint.
-func (endpoint *VhostUserEndpoint) SetProperties(properties NetworkInfo) {
+func (endpoint *VhostUserEndpoint) SetProperties(properties types.NetworkInfo) {
 	endpoint.EndpointProperties = properties
 }
 
@@ -69,12 +69,12 @@ func (endpoint *VhostUserEndpoint) SetPciAddr(pciAddr string) {
 }
 
 // NetworkPair returns the network pair of the endpoint.
-func (endpoint *VhostUserEndpoint) NetworkPair() *NetworkInterfacePair {
+func (endpoint *VhostUserEndpoint) NetworkPair() *types.NetworkInterfacePair {
 	return nil
 }
 
 // Attach for vhostuser endpoint
-func (endpoint *VhostUserEndpoint) Attach(h hypervisor.Hypervisor) error {
+func (endpoint *VhostUserEndpoint) Attach(h Hypervisor) error {
 	// Generate a unique ID to be used for hypervisor commandline fields
 	randBytes, err := utils.GenerateRandomBytes(8)
 	if err != nil {
@@ -89,7 +89,7 @@ func (endpoint *VhostUserEndpoint) Attach(h hypervisor.Hypervisor) error {
 		Type:       config.VhostUserNet,
 	}
 
-	return h.AddDevice(d, hypervisor.VhostuserDev)
+	return h.AddDevice(d, VhostuserDev)
 }
 
 // Detach for vhostuser endpoint
@@ -98,17 +98,17 @@ func (endpoint *VhostUserEndpoint) Detach(netNsCreated bool, netNsPath string) e
 }
 
 // HotAttach for vhostuser endpoint not supported yet
-func (endpoint *VhostUserEndpoint) HotAttach(h hypervisor.Hypervisor) error {
+func (endpoint *VhostUserEndpoint) HotAttach(h Hypervisor) error {
 	return fmt.Errorf("VhostUserEndpoint does not support Hot attach")
 }
 
 // HotDetach for vhostuser endpoint not supported yet
-func (endpoint *VhostUserEndpoint) HotDetach(h hypervisor.Hypervisor, netNsCreated bool, netNsPath string) error {
+func (endpoint *VhostUserEndpoint) HotDetach(h Hypervisor, netNsCreated bool, netNsPath string) error {
 	return fmt.Errorf("VhostUserEndpoint does not support Hot detach")
 }
 
 // Create a vhostuser endpoint
-func createVhostUserEndpoint(netInfo NetworkInfo, socket string) (*VhostUserEndpoint, error) {
+func createVhostUserEndpoint(netInfo types.NetworkInfo, socket string) (*VhostUserEndpoint, error) {
 
 	vhostUserEndpoint := &VhostUserEndpoint{
 		SocketPath:   socket,
@@ -121,7 +121,7 @@ func createVhostUserEndpoint(netInfo NetworkInfo, socket string) (*VhostUserEndp
 
 // findVhostUserNetSocketPath checks if an interface is a dummy placeholder
 // for a vhost-user socket, and if it is it returns the path to the socket
-func findVhostUserNetSocketPath(netInfo NetworkInfo) (string, error) {
+func findVhostUserNetSocketPath(netInfo types.NetworkInfo) (string, error) {
 	if netInfo.Iface.Name == "lo" {
 		return "", nil
 	}
@@ -143,7 +143,7 @@ func findVhostUserNetSocketPath(netInfo NetworkInfo) (string, error) {
 func vhostUserSocketPath(info interface{}) (string, error) {
 
 	switch v := info.(type) {
-	case NetworkInfo:
+	case types.NetworkInfo:
 		return findVhostUserNetSocketPath(v)
 	default:
 		return "", nil
